requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
pytz
requests
Flask-WTF
passlib
flask_basicauth
psycopg2-binary==2.9.1
